Move Windows-only resource files to Platform/Windows/ (#343)
Group Windows-specific resource files into src/source/Platform/Windows/
so the source root no longer mixes platform-specific files with
cross-platform code. Sets up the conventional Platform/<os>/ layout
used by Godot, Qt, SDL.

Moved:
- resource.h, resource.rc, resource.aps
- icon1.ico, icon2.ico
- Global Release/resource.res (also drops the stray-space folder)

Updated:
- Winmain.cpp: #include "Platform/Windows/resource.h"
- src/CMakeLists.txt: target_sources path for the RC file
