+++ src/CMakeLists.txt
@@ -141,7 +141,7 @@ if (CMAKE_SYSTEM_NAME STREQUAL "Windows")
 
   enable_language(RC)
   target_sources(Main PRIVATE
-    "${CMAKE_CURRENT_SOURCE_DIR}/source/resource.rc"
+    "${CMAKE_CURRENT_SOURCE_DIR}/source/Platform/Windows/resource.rc"
   )
 endif()
