+++ src/source/Platform/Windows/icon2.ico
