+++ src/source/Platform/Windows/resource.aps
