+++ src/source/Platform/Windows/resource.h
