+++ src/source/Platform/Windows/resource.rc
