+++ src/source/Platform/Windows/resource.res
